sunxi: Store the device tree name in the SPL header
This patch updates the mksunxiboot tool to optionally add
the default device tree name string to the SPL header. This
information can be used by the firmware upgrade tools to
protect users from harming themselves by trying to upgrade
to an incompatible bootloader.

The primary use case here is a non-removable bootable media
(such as NAND, eMMC or SPI flash), which already may have
a properly working, but a little bit outdated bootloader
installed. For example, the user may download or build a
new U-Boot image for "Cubieboard", and then attemept to
install it on a "Cubieboard2" hardware by mistake as a
replacement for the already existing bootloader. If this
happens, the flash programming tool can identify this
problem and warn the user.

CONFIG_OF_LIST may provide more than one compatible device
tree name when supporting multiple boards by a single
defconfig. In such cases, the right device tree name is
getting identified at runtime from the list of multiple
options. The patch calculates a 32-bit hash value (crc32)
of this list and stores it to the SPL header too, because
this is better than nothing and still can be used to warn
about potentially incompatible upgrades.

The size of the SPL header is also increased from 64 bytes
to 96 bytes to provide enough space for the device tree name
string.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>
Signed-off-by: Andre Przywara <andre.przywara@arm.com>